warnings removed (#454)
bison and flex code generation move to build
some tests renamed for easier navigation
example folder added to ubuntu dockerfile for easier testing
improved includes
improved inheritance for some classes


* no more warnings

* fix for rebase errors

* Bison/Flex generation

* example added to docker-ubuntu

* clang warnings

* release build warnings

* tests renaming for easier navigation

* fix dockerfile examples

* includes and inheritance

* flex and bison source files formating

* git actions update with temp removal of -Werror

* macos warnings

* block and segment_tree header fields unified

* pg_ptr_assert_cast fix

* remaining warnings

* to_pmr_string fix

* turn on -Werror except sanitizers and macOS

* -Werror flag cleanup

+++ components/catalog/versioned_trie/versioned_trie_node.hpp
@@ -171,9 +171,9 @@ namespace components::catalog {
         void erase(std::size_t i) {
             // this empty-keys situation happens only in the header node.
             if (!keys_.empty()) {
-                keys_.erase(keys_.begin() + i);
+                keys_.erase(keys_.begin() + static_cast<int64_t>(i));
             }
-            auto it = children_.erase(children_.begin() + i);
+            auto it = children_.erase(children_.begin() + static_cast<int64_t>(i));
             parent_index_.erase(it, children_.end());
         }
 
@@ -200,7 +200,7 @@ namespace components::catalog {
 
     private:
         versioned_trie_node const* const_this() { return const_cast<versioned_trie_node const*>(this); }
-        key_element const& key(const_iterator it) const { return keys_[it - children_.begin()]; }
+        key_element const& key(const_iterator it) const { return keys_[static_cast<size_t>(it - children_.begin())]; }
 
         keys_t keys_;
         children_t children_;
