+++ components/tests/generaty.cpp
@@ -88,11 +88,11 @@ components::vector::data_chunk_t gen_data_chunk(size_t size, int num, std::pmr::
     chunk.set_cardinality(size);
 
     for (size_t i = 1; i <= size; i++) {
-        chunk.set_value(0, i - 1, logical_value_t{static_cast<int64_t>(i + num)});
-        chunk.set_value(1, i - 1, logical_value_t{gen_id(i + num)});
-        chunk.set_value(2, i - 1, logical_value_t{std::to_string(i + num)});
-        chunk.set_value(3, i - 1, logical_value_t{double(i + num) + 0.1});
-        chunk.set_value(4, i - 1, logical_value_t{(i + num) % 2 != 0});
+        chunk.set_value(0, i - 1, logical_value_t{static_cast<int64_t>(i) + num});
+        chunk.set_value(1, i - 1, logical_value_t{gen_id(static_cast<int>(i) + num)});
+        chunk.set_value(2, i - 1, logical_value_t{std::to_string(static_cast<int64_t>(i) + num)});
+        chunk.set_value(3, i - 1, logical_value_t{static_cast<double>(static_cast<int64_t>(i) + num) + 0.1});
+        chunk.set_value(4, i - 1, logical_value_t{(static_cast<int64_t>(i) + num) % 2 != 0});
         {
             std::vector<logical_value_t> arr;
             arr.reserve(array_size);
